+++ OrderService/Models/OrderItem.cs
@@ -24,7 +24,7 @@ namespace OrderService.Models
         public int quantity { get; set; }
 
         // Refers to cost of 1 item, not many
-        public double cost { get; set; }
+        public decimal cost { get; set; }
 
         public Boolean active { get; set; }
     }
